pyproject.toml
[project]
name = "assistant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dotenv>=0.9.9",
    "evaluate>=0.4.6",
    "faiss-cpu>=1.13.2",
    "fastapi>=0.128.0",
    "google-genai>=1.59.0",
    "numpy>=2.0.0",
    "uvicorn>=0.40.0",
]
